The Mulder Files
This is the software for Mulder, an animatronic skull. Number 6 in the
Mike's Pi Bakery series:-
published in The MagPi No. 38 October 2015

Note as a bonus there is also Mulder Script, a scripting engine that
can drive Mulder through a pre defined sequence of movements, sound and
light changes. See the notes on how to drive it. The example "Script1"
was used to make the movie at
https://vimeo.com/139202162