Latest release of TAP-A-LED applications
Video at  https://youtu.be/Cc29lIrQY8I
Part 2 video shows some of the applications written for the TAP-A-LED
hardware. This is published in The MagPi magazine number 98.
It includes:-
TAP-A-Sketch - Drawing on the display
TAP-AN-Animation - Sequence of screens shown in sequence
Lights Off - Try and get all the LEDs, but as one turns off others
invert, the setup screens can be a lot harder that this example.
TAP-A-Sequence - A 16 step sound sequence generator both MIDI and
sample.